rename: feature 폴더에 이벤트 생성과 관련 event-create 폴더로 분리
