+++ src/features/event-manage/event-create/ui/LinkInput.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import AddButton from '../../../../public/assets/event-manage/creation/AddBtn.svg';
-import CloseButton from '../../../../public/assets/event-manage/creation/CloseBtn.svg';
-import Link from '../../../../public/assets/event-manage/creation/Link.svg';
+import AddButton from '../../../../../public/assets/event-manage/creation/AddBtn.svg';
+import CloseButton from '../../../../../public/assets/event-manage/creation/CloseBtn.svg';
+import Link from '../../../../../public/assets/event-manage/creation/Link.svg';
 import { FunnelState } from '../model/FunnelContext';
 
 interface LinkInputProps {
