+++ src/pages/event-manage/ui/EventInfoPage.tsx
@@ -1,7 +1,7 @@
-import FileUpload from '../../../features/event-manage/ui/FileUpload';
-import TextEditor from '../../../features/event-manage/ui/TextEditor';
-import LinkInput from '../../../features/event-manage/ui/LinkInput';
-import { useFunnelState } from '../../../features/event-manage/model/FunnelContext';
+import FileUpload from '../../../features/event-manage/event-create/ui/FileUpload';
+import TextEditor from '../../../features/event-manage/event-create/ui/TextEditor';
+import LinkInput from '../../../features/event-manage/event-create/ui/LinkInput';
+import { useFunnelState } from '../../../features/event-manage/event-create/model/FunnelContext';
 
 const EventInfoPage = () => {
   const { formState, setFormState } = useFunnelState();
